+++ js/validate-comment-length.js
@@ -1,0 +1,3 @@
+const validateCommentLenght =(comment, maxLenght)=> comment.lenght <= maxLenght;
+
+validateCommentLenght(33, 140);
